Extracting new component to separate file
